Adding tracking/emotion classification to the app
Using the modified versions functions of trackerWrapper, a basic
version of tracking/classification is now available whilst viewing
experiments. At the moment, it will log the current item you’re looking
at and then the confidence values for the 8 classes for each frame.

A future commit will then take this data, put into a JSON format and
send back to the API to view on the experiments web app